requirements.txt
################################################################
# TensorFlow can be installed either as CPU or GPU versions.
# You select which one to install by (un)commenting these lines.

# tensorflow		# CPU Version of TensorFlow.
 tensorflow-gpu	# GPU version of TensorFlow.

################################################################
# Other libraries:

numpy
matplotlib
pillow

# Face crop:
mtcnn

################################################################
